@@ -16,7 +16,7 @@ flutter:
         default_package: video_player_web
 
 dependencies:
-  meta: 1.3.0
+  meta: 1.7.0
   video_player_platform_interface: 4.1.0
 
   # The design on https://flutter.dev/go/federated-plugins was to leave
@@ -30,14 +30,21 @@ dependencies:
     sdk: flutter
   
   http: ^0.13.3
-  youtube_explode_dart: ^1.9.4
+  html: ^0.15.0
+  xml: ^5.0.2
+  freezed_annotation: ^0.14.3
+  json_annotation: 4.1.0
+  #youtube_explode_dart:
+  #  git:
+  #    url: https://github.com/DimaBrahin/youtube_explode_dart
+  #    ref: urls_obtain_order
+  #youtube_explode_dart: ^1.9.4
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
   pedantic: ^1.8.0
   pigeon: 0.1.7
+  build_runner: ^1.3.1
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.14.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.5 <2.0.0"
